pyproject.toml
[build-system]
requires = ["setuptools>=64.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "visarchpy"
version = "0.4.0"
authors = [
    {name = "Manuel Garcia", email = "m.g.garciaalvarez@tudelft.nl"},
]
description = "data pipelines for extraction of architectural visuals"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["one", "two"]
license = {text = "BSD-3-Clause"}
classifiers = ["Development Status :: 3 - Alpha",
 "Environment :: Python",
 "Intended Audience :: Science/Research",
 "Operating System :: OS Independent",
 "Programming Language :: Python :: 3",
]
dependencies = [
    "Pillow",
    "requests",
    "pdfminer.six[image]",
    "beautifulsoup4",
    "shapely",
    "pandas",
    "pymods",
    "tqdm",
    "pdf2image",
    "pytesseract",
    "matplotlib",
    "typer",
    "transformers",
    "scikit-learn",
]

[tool.pytest.ini_options]
pythonpath = "src"
addopts = [
    "--import-mode=importlib",
]

[project.scripts]
visarch = "visarchpy.cli:app"

